requirements.txt
torch
torchvision
Pillow==9.1.1
scikit-image==0.19.3  
scikit-learn==1.1.2
sklearn==0.0
opencv-python==4.6.0.66
grad_cam==1.4.3   
tqdm==4.61.2
PyYAML==6.0
easydict==1.9
ftfy==6.1.3
ftfy==6.1.3
regex==2023.12.25
imgaug==0.4.0
numpy==1.22.4
